Consider existing fridge stock before making purchase with buyListedItems. Add the loop for consuming food

+++ src/main/scala/meta/example/supermarket/utils/CategoryAmount.scala
@@ -1,10 +1,10 @@
 package meta.example.supermarket
 
-case class categoryAmount(var Vegetable: Double=0.0,
-                          var Meat: Double=0.0,
-                          var Snack: Double=0.0,
-                          var Grain: Double=0.0,
-                          var Dairy: Double=0.0) {
+case class categoryAmount(var Vegetable: Int = 0,
+                          var Meat: Int = 0,
+                          var Snack: Int =0,
+                          var Grain: Int =0,
+                          var Dairy: Int =0) {
   def +(bar: categoryAmount): categoryAmount = {
     categoryAmount(bar.Vegetable + Vegetable, bar.Meat+Meat,
       bar.Snack + Snack, bar.Grain + Grain, bar.Dairy + Dairy)
